+++ fiddles/angular2-cli/fiddle-0030-EntAgGridLinearRegression/src/app/app.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms'; // <-- NgModel lives here
+import { HttpClientModule } from '@angular/common/http';
+import { AgGridModule } from 'ag-grid-angular';
+import { AppComponent } from './app.component';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    FormsModule, // <-- import the FormsModule before binding with [(ngModel)]
+    HttpClientModule,
+    AgGridModule.withComponents([])
+  ],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }