+++ CHANGELOG.md
@@ -2184,3 +2184,23 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 
   1.  Added [fiddles/java/fiddle-0006-j2v8RestApi](fiddles/java/fiddle-0006-j2v8RestApi)
 
+
+### 320_AngularCli_28 ~ [Issue 320](https://github.com/bradyhouse/house/issues/320)
+
+* **Angular-Cli**
+
+  1. Added [fiddles/angular2-cli/fiddle-0028-EntAgGridBarChart](fiddles/angular2-cli/fiddle-0028-EntAgGridBarChart)
+
+
+### 322_AngularCli_29 ~ [Issue 322](https://github.com/bradyhouse/house/issues/322)
+
+* **Angular-Cli**
+
+  1. Added [fiddles/angular2-cli/fiddle-0029-EntAgGridChartSeriesClick](fiddles/angular2-cli/fiddle-0029-EntAgGridChartSeriesClick)
+
+
+### 324_AngularCli_30 ~ [Issue 324](https://github.com/bradyhouse/house/issues/324)
+
+* **Angular-Cli**
+
+    1. Added [fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect](fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect)